requirements.txt
# Bread SDK
aibread>=1.0.0

# Dependencies
python-dotenv>=1.0.0
requests>=2.31.0

